Move Elevation View rotation controls into its own view.
